Get TypeScript set up.
