+++ web-playground/jest.config.js
@@ -1,5 +1,8 @@
 module.exports = {
     preset: "jest-puppeteer",
+    transform: {
+        "^.+\\.ts?$": "ts-jest"
+    },
     globals: {
         URL: "http://localhost:3000/playground"
     },
