+++ web-playground/package.json
@@ -13,22 +13,23 @@
     "@testing-library/react": "^14.0.0",
     "@testing-library/user-event": "^14.4.3",
     "@types/expect-puppeteer": "^5.0.3",
+    "@types/jest": "^29.5.0",
     "@types/jest-environment-puppeteer": "^5.0.3",
-    "@types/puppeteer": "^7.0.4",
     "@types/react": "^18.0.30",
     "@types/react-dom": "^18.0.11",
     "jest": "^29.5.0",
     "jest-environment-jsdom": "^29.5.0",
     "jest-puppeteer": "^6.2.0",
     "puppeteer": "^18.2.1",
-    "react-scripts": "^5.0.1"
+    "react-scripts": "^5.0.1",
+    "ts-jest": "^29.0.5"
   },
   "scripts": {
     "start": "react-scripts start",
     "build": "react-scripts build",
     "test": "react-scripts test",
     "eject": "react-scripts eject",
-    "e2e": "cd e2e && jest"
+    "e2e": "jest"
   },
   "eslintConfig": {
     "extends": [
